Move useEventOver to WeatherEvent
