+++ src/features/game/expansion/components/temperateSeason/WeatherEvent.tsx
@@ -7,8 +7,7 @@ import {
   NoticeboardItemsElements,
 } from "features/world/ui/kingdom/KingdomNoticeboard";
 import { useAppTranslation } from "lib/i18n/useAppTranslations";
-import React, { useCallback, useRef, useState } from "react";
-import { useEventOver } from "./CalendarEvent";
+import React, { useCallback, useEffect, useRef, useState } from "react";
 import { SeasonalEventName } from "features/game/types/calendar";
 import { useSelector } from "@xstate/react";
 import { useGame } from "features/game/GameProvider";
@@ -23,6 +22,33 @@ interface Props {
   acknowledge: () => void;
   handleAFK: () => void;
   showEventIcons?: boolean;
+}
+
+function useEventOver({ setEventOver }: { setEventOver: () => void }) {
+  // Calculate time until next check
+  const getNextCheckTime = () => {
+    const now = new Date();
+
+    // In development: check at the start of each minute
+    // const nextMinute = new Date(now);
+    // nextMinute.setSeconds(0);
+    // nextMinute.setMilliseconds(0);
+    // nextMinute.setMinutes(nextMinute.getMinutes() + 1);
+    // return nextMinute.getTime() - now.getTime();
+
+    // In production: check at UTC midnight
+    const tomorrow = new Date(now);
+    tomorrow.setUTCHours(24, 0, 0, 0);
+    return tomorrow.getTime() - now.getTime();
+  };
+
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      setEventOver();
+    }, getNextCheckTime());
+
+    return () => clearTimeout(timeout);
+  }, [setEventOver]);
 }
 
 export const WeatherEvent: React.FC<Props> = ({
@@ -58,7 +84,12 @@ export const WeatherEvent: React.FC<Props> = ({
     })),
   );
 
-  const getEventEndTime = (eventStartTime: number) => {
+  const eventStartTime = useSelector(
+    gameService,
+    (state: MachineState) =>
+      state.context.state.calendar[eventName]?.triggeredAt,
+  );
+  const getEventEndTime = () => {
     // In development: check at the start of each minute
     // const now = new Date();
     // const nextMinute = new Date(now);
@@ -68,18 +99,12 @@ export const WeatherEvent: React.FC<Props> = ({
     // return nextMinute.getTime();
 
     // In production: check at UTC midnight
-    const tomorrow = new Date(eventStartTime);
+    const tomorrow = new Date(eventStartTime ?? Date.now());
     tomorrow.setUTCHours(24, 0, 0, 0);
     return tomorrow.getTime();
   };
 
-  const eventStartTime = useSelector(
-    gameService,
-    (state: MachineState) =>
-      state.context.state.calendar[eventName]?.triggeredAt,
-  );
-
-  const eventEndTime = getEventEndTime(eventStartTime ?? Date.now());
+  const eventEndTime = getEventEndTime();
 
   return (
     <>
